auto-pde-agent/src/autopdeagent/__init__.py
"""
This file provides:

- Path settings for global config file & relative directories
- Version numbering
- Protocols for the core components of autopdeagent.
  By the magic of protocols & duck typing, you can pretty much ignore them,
  unless you want the static type checking.
"""

# 💡 建议：既然是你们自己的开源项目，可以把版本号改成 0.1.0 或 1.0.0
__version__ = "1.0.0"

import os
from pathlib import Path
from typing import Any, Protocol

import dotenv
from platformdirs import user_config_dir
from rich.console import Console

# === 🔄 修改 1：替换包名为 autopdeagent ===
from autopdeagent.utils.log import logger

package_dir = Path(__file__).resolve().parent

# === 🔄 修改 2：替换环境变量和本地配置文件夹名 ===
global_config_dir = Path(os.getenv("AUTOPDE_GLOBAL_CONFIG_DIR") or user_config_dir("autopdeagent"))
global_config_dir.mkdir(parents=True, exist_ok=True)
global_config_file = Path(global_config_dir) / ".env"

# === 🔄 修改 3：替换静默启动环境变量和终端欢迎语 ===
if not os.getenv("AUTOPDE_SILENT_STARTUP"):
    Console().print(
        f" This is [bold green]AutoPDEAgent[/bold green] version [bold green]{__version__}[/bold green].\n"
        f"Loading global config from [bold green]'{global_config_file}'[/bold green]"
    )
dotenv.load_dotenv(dotenv_path=global_config_file)


# === Protocols ===
# You can ignore them unless you want static type checking.

class Model(Protocol):
    """Protocol for language models."""

    config: Any
    cost: float
    n_calls: int

    def query(self, messages: list[dict[str, str]], **kwargs) -> dict: ...

    def get_template_vars(self) -> dict[str, Any]: ...


class Environment(Protocol):
    """Protocol for execution environments."""

    config: Any

    def execute(self, command: str, cwd: str = "") -> dict[str, str]: ...

    def get_template_vars(self) -> dict[str, Any]: ...


class Agent(Protocol):
    """Protocol for agents."""

    model: Model
    env: Environment
    messages: list[dict[str, str]]
    config: Any

    def run(self, task: str, **kwargs) -> tuple[str, str]: ...


__all__ = [
    "Agent",
    "Model",
    "Environment",
    "package_dir",
    "__version__",
    "global_config_file",
    "global_config_dir",
    "logger",
]